# Conflicts:
#	btc_indexer/crates/config_parser/src/config.rs
#	gateway/crates/config_parser/src/config.rs
#	gateway/crates/flow_processor/src/flow_router.rs
#	gateway/crates/flow_processor/src/types.rs
#	gateway/crates/local_db_store/src/schemas/musig_identifier.rs
#	gateway/crates/local_db_store/src/schemas/sign_session.rs
#	gateway/crates/local_db_store/src/storage.rs
#	verifier/apps/main/src/main.rs
#	verifier/crates/local_db_store/src/schemas/musig_identifier.rs
#	verifier/crates/local_db_store/src/schemas/sign_session.rs
#	verifier/crates/local_db_store/src/storage.rs

+++ gateway/crates/local_db_store/src/schemas/deposit_address.rs
@@ -21,7 +21,7 @@ pub struct DepositAddrInfo {
     pub nonce_tweak: Vec<u8>,
     pub address: String,
     pub is_btc: bool,
-    pub amount: i32,
+    pub amount: u64,
     pub confirmation_status: DepositStatus,
 }
